+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -7,6 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationResponse;
+import roomescape.dto.SaveReservationRequest;
 
 import java.util.ArrayList;
 import java.util.List;
