+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,4 +1,6 @@
-package roomescape;
+package roomescape.dto;
+
+import roomescape.domain.Reservation;
 
 import java.util.List;
