+++ src/main/java/roomescape/dto/SaveReservationRequest.java
@@ -1,4 +1,6 @@
-package roomescape;
+package roomescape.dto;
+
+import roomescape.domain.Reservation;
 
 public record SaveReservationRequest(String name, String date, String time) {
     public static Reservation toEntity(Long id, SaveReservationRequest request) {
